+++ front/src/pages/WriteReview.tsx
@@ -1,5 +1,5 @@
 import PageTemplate from "@/shared/original-ui/PageTemplate";
-import ReviewForm from "@/features/review/ReviewForm";
+import ReviewForm from "@/widgets/ReviewForm";
 
 export default function WriteReview() {
   return (
